+++ src/publisher/browser/index.ts
@@ -1,9 +1,13 @@
-import { GetRemoteUIHandler, Options } from '../types.js';
+import { GetRemoteUIHandler, PublisherOptions } from '../../types.js';
 import EventTransport from '../../transport/event.js';
 import { getPublisher, connect } from '../factory.js';
 import { createBrowserWsTransport, fetchWsSettings } from './transport-ws.js';
 
-export function createPublisher(id: string, getRemoteUI: GetRemoteUIHandler, options?: Options) {
+export function createPublisher(
+    id: string,
+    getRemoteUI: GetRemoteUIHandler,
+    options?: PublisherOptions
+) {
     connect(true, createBrowserWsTransport, fetchWsSettings);
 
     const publisher = getPublisher(id, getRemoteUI, options);
@@ -19,7 +23,9 @@ export function createPublisher(id: string, getRemoteUI: GetRemoteUIHandler, opt
     // self subscriber
     EventTransport.get('rempl-self-publisher', 'rempl-self-subscriber').sync(publisher);
 
-    return publisher;
+    return Object.assign(publisher.ns('*'), {
+        ns: publisher.ns.bind(publisher),
+    });
 }
 
 export function connectPublisherWs(uri?: string) {